moved required files to its own file and System Module. Added OS check for safe messure
