+++ lib/lib.rb
@@ -1,0 +1,21 @@
+#
+#
+
+# gems 
+require 'colorize'
+require 'open3'
+require 'pty'
+require 'json'
+require 'net/http'
+
+
+
+# local files 
+require 'utility'
+require 'rvm'
+require 'brew'
+require 'ruby'
+require 'osx'
+require 'zsh'
+require 'rbenv'
+
